chatbot/app/requirements.txt
streamlit
boto3
streamlit-cognito-auth
pandas
PyPDF2
python-docx
